+++ test/Unit/Parser/MariadbParserTest.php
@@ -5,20 +5,20 @@ declare(strict_types=1);
 namespace Pongee\DatabaseSchemaVisualization\Test\Unit\Parser;
 
 use Override;
-use Pongee\DatabaseSchemaVisualization\Parser\MysqlParser;
+use Pongee\DatabaseSchemaVisualization\Parser\MariadbParser;
 use Pongee\DatabaseSchemaVisualization\Parser\ParserInterface;
 
-final class MysqlParserTest extends FullSchemaParserTestCase
+final class MariadbParserTest extends FullSchemaParserTestCase
 {
     #[Override]
     protected static function fixturesSubDirectory(): string
     {
-        return '/Mysql/';
+        return '/Mariadb/';
     }
 
     #[Override]
     protected function createParser(): ParserInterface
     {
-        return new MysqlParser();
+        return new MariadbParser();
     }
 }